requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
config==0.5.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
praw==7.6.1
prawcore==2.3.0
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
